requirements.txt
pywebio
psycopg2-binary

